requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==25.4.0
boto3==1.42.88
botocore==1.42.88
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
dill==0.4.1
docker==7.1.0
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lightning-utilities==0.15.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
packaging==24.2
pandas==3.0.2
pathos==0.3.5
pillow==12.2.0
platformdirs==4.9.6
pox==0.3.7
ppft==1.7.8
pretty-errors==1.2.25
protobuf==6.33.6
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==14.3.4
rpds-py==0.30.0
s3transfer==0.16.0
sagemaker==2.257.1
sagemaker-core==1.0.77
schema==0.7.8
setuptools==81.0.0
six==1.17.0
smdebug-rulesconfig==1.0.1
starlette==1.0.0
sympy==1.14.0
tblib==3.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.11.0+cu130
torchmetrics==1.4.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.42.0
Werkzeug==3.1.8
zipp==3.23.0
